Add GetQueueUrl action.
This adds the GetQueueUrl action to Vassal, along with an integration
test to check that it works.  Since this means we can actually check
if queues have been created, this also includes some fixes for various
bits of queue creation.

Also includes the first bit of error reporting work - throwing or
returning a Vassal.Results.SQSError should cause the WebRouter to return
some error XML.