Add Queue.Receiver for managing receives on a Queue.
This commit adds Queue.Receiver, which will manage receives on a queue.
When a receive initially comes in, we'll check if there's anything to
send to it.  If not, we add it to a list of pending receives.

When there are pending receives, we'll check every 200ms for new
messages.  If there are no new messages, we will wait, otherwise we will
fufil the request.

Clients need to acknowledge receipt of the messages with the Receiver,
otherwise it will assume they have timedout and re-add the messages into
the queue.  This acknowledgement is dealt with in the Receiver API, so
other code should not need to worry about it.

This is just the first pass for the receiver - when queues are quiet we
will end up with extremely long lists of pending receives, most of which
will have timed out.  This can be addressed in a future commit.