ReceiveRequests can now be cancelled in Receiver.
This fixes a potential problem with the Receiver code - processes that
have timed out can now cancel their requests, rather than the receiver
just assuming they are still around and waiting for a timeout after
sending on the messages.

There is still a potential race condition between request cancellation
and messages being sent, but it's a reduced risk - and the messages
should never be acked, so messages should not be lost.