Remove QueueManager, move remaining actions into Queue.
Decided that having a QueueManager process didn't make sense - much like
the centralised Queue process, it's just an un-necessary bottleneck.
Also, having just 2 actions go into the QueueManager, while the rest go
into the Queue felt a bit arbitrary.

So, I've ditched all the code for the QueueManager, and moved the code
that handled it's actions into the Queue module.