Don't use :timer.send_interval.
While looking into things that could be causing #7 I noticed that I'm
using the :timer module.  According to [the erlang common
caveats](http://erlang.org/doc/efficiency_guide/commoncaveats.html) I
should not use the functions in the :timer module.

I'm not sure if this is the cause of #7, but I'm going to see if it
helps.