Travis now builds & upload ubuntu releases.
Building ubuntu releases has been a bit of a pain in the past, as I'm
running mac OS.  It involved shelling into an ubuntu VM and running the
build there, then uploading some files manually.

This automates the ubuntu release building on travis CI.  When a tag is
pushed up, it will build a release and upload it to github
automatically.

Eventually we might be able to extend this to do the OS X release as
well, but the elixir support on travis mac os images seems patchy right
now.

Fixes #8.