Return as quick as possible when no wait time.
When making a ReceiveMessages request, if a client doesn't provide a
WaitTimeSeconds, then we should return to them immediately if there are no
messages.

Previously the Receiver always handled requests asyncrhonously, so we introduced
an artificial wait time of 500ms when none was provided.

This commit updates the Receiver with a syncrhonous code path for when these
requests are made.